Re-add support for dryrun

+++ src/upload.rs
@@ -9,7 +9,7 @@ use symbolic_debuginfo::FileFormat;
 use crate::config;
 use crate::symstore;
 
-pub fn upload(search_path: &Path, recursive: bool, server: &config::RemoteStorage) -> Result<()> {
+pub fn upload(search_path: &Path, recursive: bool, server: &config::RemoteStorage, dryrun: bool) -> Result<()> {
     let files = find_obj_files(search_path, recursive)?;
     let files = map_files_to_keys(&files);
     match &server.storage_type {
@@ -17,9 +17,9 @@ pub fn upload(search_path: &Path, recursive: bool, server: &config::RemoteStorag
             "Upload to HTTP server ({}) not yet implemented!",
             c.url
         )),
-        config::RemoteStorageType::S3(c) => upload_to_s3(c, &files),
-        config::RemoteStorageType::B2(c) => upload_to_b2(c, &files),
-        config::RemoteStorageType::Path(c) => copy_to_folder(c, &files),
+        config::RemoteStorageType::S3(c) => upload_to_s3(c, &files, dryrun),
+        config::RemoteStorageType::B2(c) => upload_to_b2(c, &files, dryrun),
+        config::RemoteStorageType::Path(c) => copy_to_folder(c, &files, dryrun),
     }
 }
 
@@ -91,7 +91,7 @@ fn map_files_to_keys(files: &Vec<PathBuf>) -> Vec<(PathBuf, String)> {
     map
 }
 
-fn upload_to_s3(config: &config::S3Config, files: &Vec<(PathBuf, String)>) -> Result<()> {
+fn upload_to_s3(config: &config::S3Config, files: &Vec<(PathBuf, String)>, dryrun: bool) -> Result<()> {
     let creds = s3::creds::Credentials::new(None, None, None, None, config.profile.as_deref())?;
     let region = config.region.parse()?;
     let bucket = s3::bucket::Bucket::new(&config.bucket, region, creds)?;
@@ -105,13 +105,15 @@ fn upload_to_s3(config: &config::S3Config, files: &Vec<(PathBuf, String)>) -> Re
             config.bucket,
             full_key
         );
-        bucket.put_object_stream_blocking(&file.0, full_key)?;
+        if !dryrun {
+            bucket.put_object_stream_blocking(&file.0, full_key)?;
+        }
     }
 
     Ok(())
 }
 
-fn upload_to_b2(config: &config::B2Config, files: &Vec<(PathBuf, String)>) -> Result<()> {
+fn upload_to_b2(config: &config::B2Config, files: &Vec<(PathBuf, String)>, dryrun: bool) -> Result<()> {
     let b2_creds = match &config.account_id {
         Some(id) => b2creds::Credentials::from_file(None, Some(&id))?,
         None => b2creds::Credentials::default()?,
@@ -140,13 +142,15 @@ fn upload_to_b2(config: &config::B2Config, files: &Vec<(PathBuf, String)>) -> Re
             config.bucket,
             full_key
         );
-        bucket.put_object_stream_blocking(&file.0, full_key)?;
+        if !dryrun {
+            bucket.put_object_stream_blocking(&file.0, full_key)?;
+        }
     }
 
     Ok(())
 }
 
-fn copy_to_folder(config: &config::PathConfig, files: &Vec<(PathBuf, String)>) -> Result<()> {
+fn copy_to_folder(config: &config::PathConfig, files: &Vec<(PathBuf, String)>, dryrun: bool) -> Result<()> {
     for file in files {
         let dest = config.path.join(&file.1);
         if let Some(parent) = dest.parent() {
@@ -156,11 +160,13 @@ fn copy_to_folder(config: &config::PathConfig, files: &Vec<(PathBuf, String)>) -
             ))?;
         }
         println!("Copying '{}' to '{}'", file.0.display(), dest.display());
-        std::fs::copy(&file.0, &dest).context(format!(
-            "Failed to copy '{}' to '{}'",
-            file.0.display(),
-            dest.display()
-        ))?;
+        if !dryrun {
+            std::fs::copy(&file.0, &dest).context(format!(
+                "Failed to copy '{}' to '{}'",
+                file.0.display(),
+                dest.display()
+            ))?;
+        }
     }
 
     Ok(())
